pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llmace"
version = "0.1.0"
description = "Agentic Context Engineering - Evolving LLM contexts through reflection and curation"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "LLMACE Contributors", email = "llmace@example.com"}
]
maintainers = [
    {name = "LLMACE Contributors", email = "llmace@example.com"}
]
keywords = ["llm", "ai", "agents", "context", "prompt-engineering", "machine-learning", "agentic", "reflection", "curation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Operating System :: OS Independent",
    "Typing :: Typed",
]

dependencies = [
    "openai>=1.0.0",
    "pydantic>=2.0.0",
    "numpy>=1.20.0",
    "python-dotenv>=0.9.9",
    "backoff>=2.0.0",
]

[project.optional-dependencies]
embeddings = [
    "sentence-transformers>=2.0.0",
]
agents = [
    "langgraph>=0.0.20",
    "langchain-openai>=0.0.5",
    "langchain-core>=0.1.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]
all = [
    "sentence-transformers>=2.0.0",
    "langgraph>=0.0.20",
    "langchain-openai>=0.0.5",
    "langchain-core>=0.1.0",
]

[project.urls]
Homepage = "https://github.com/llmace/llmace"
Documentation = "https://github.com/llmace/llmace#readme"
Repository = "https://github.com/llmace/llmace"
Issues = "https://github.com/llmace/llmace/issues"
Changelog = "https://github.com/llmace/llmace/blob/main/CHANGELOG.md"
"Source Code" = "https://github.com/llmace/llmace"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ['py39']

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
